Move Byron code to `Ouroboros.Consensus.Byron.ByronHFC`
