+++ ouroboros-consensus-cardano/src/ouroboros-consensus-cardano/Ouroboros/Consensus/Cardano/CanHardFork.hs
@@ -1,7 +1,5 @@
 {-# LANGUAGE ConstraintKinds #-}
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE DeriveAnyClass #-}
-{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE EmptyCase #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE FlexibleInstances #-}
@@ -9,7 +7,6 @@
 {-# LANGUAGE MultiParamTypeClasses #-}
 {-# LANGUAGE NamedFieldPuns #-}
 {-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE RecordWildCards #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE TypeApplications #-}
 {-# LANGUAGE TypeFamilies #-}
@@ -19,8 +16,7 @@
 {-# OPTIONS_GHC -Wno-orphans #-}
 
 module Ouroboros.Consensus.Cardano.CanHardFork (
-    ByronPartialLedgerConfig (..)
-  , CardanoHardForkConstraints
+    CardanoHardForkConstraints
   , TriggerHardFork (..)
     -- * Re-exports of Shelley code
   , ShelleyPartialLedgerConfig (..)
@@ -28,9 +24,6 @@ module Ouroboros.Consensus.Cardano.CanHardFork (
   , translateChainDepStateAcrossShelley
   ) where
 
-import qualified Cardano.Chain.Common as CC
-import qualified Cardano.Chain.Genesis as CC.Genesis
-import qualified Cardano.Chain.Update as CC.Update
 import           Cardano.Crypto.DSIGN (Ed25519DSIGN)
 import           Cardano.Crypto.Hash.Blake2b (Blake2b_224, Blake2b_256)
 import qualified Cardano.Ledger.Core as SL
@@ -44,23 +37,18 @@ import           Cardano.Ledger.Shelley.Translation
 import qualified Cardano.Protocol.TPraos.API as SL
 import qualified Cardano.Protocol.TPraos.Rules.Prtcl as SL
 import qualified Cardano.Protocol.TPraos.Rules.Tickn as SL
-import           Control.Monad
 import           Control.Monad.Except (runExcept, throwError)
 import           Data.Coerce (coerce)
 import qualified Data.Map.Strict as Map
-import           Data.Maybe (listToMaybe, mapMaybe)
 import           Data.Proxy
 import           Data.SOP.BasicFunctors
 import           Data.SOP.InPairs (RequiringBoth (..), ignoringBoth)
 import qualified Data.SOP.Strict as SOP
 import           Data.SOP.Tails (Tails (..))
 import qualified Data.SOP.Tails as Tails
-import           Data.Word
-import           GHC.Generics (Generic)
-import           NoThunks.Class (NoThunks)
 import           Ouroboros.Consensus.Block
+import           Ouroboros.Consensus.Byron.ByronHFC ()
 import           Ouroboros.Consensus.Byron.Ledger
-import qualified Ouroboros.Consensus.Byron.Ledger.Inspect as Byron.Inspect
 import           Ouroboros.Consensus.Byron.Node ()
 import           Ouroboros.Consensus.Cardano.Block
 import           Ouroboros.Consensus.Forecast
@@ -74,8 +62,8 @@ import           Ouroboros.Consensus.Ledger.SupportsMempool (ByteSize32,
                      IgnoringOverflow, TxMeasure)
 import           Ouroboros.Consensus.Ledger.SupportsProtocol
                      (LedgerSupportsProtocol)
-import           Ouroboros.Consensus.Protocol.Abstract
-import           Ouroboros.Consensus.Protocol.PBFT (PBft, PBftCrypto)
+import           Ouroboros.Consensus.Protocol.Abstract (SelectView,
+                     TranslateProto)
 import           Ouroboros.Consensus.Protocol.PBFT.State (PBftState)
 import qualified Ouroboros.Consensus.Protocol.PBFT.State as PBftState
 import           Ouroboros.Consensus.Protocol.Praos (Praos)
@@ -89,163 +77,6 @@ import           Ouroboros.Consensus.Shelley.ShelleyHFC
 import           Ouroboros.Consensus.TypeFamilyWrappers
 import           Ouroboros.Consensus.Util (eitherToMaybe)
 import           Ouroboros.Consensus.Util.RedundantConstraints
-
-{-------------------------------------------------------------------------------
-  Figure out the transition point for Byron
-
-  The Byron ledger defines the update 'State' in
-  "Cardano.Chain.Update.Validation.Interface". The critical piece of state we
-  need is
-
-  > candidateProtocolUpdates :: ![CandidateProtocolUpdate]
-
-  which are the update proposals that have been voted on, accepted, and
-  endorsed, and now need to become stable. In `tryBumpVersion`
-  ("Cardano.Chain.Update.Validation.Interface.ProtocolVersionBump") we
-  find the candidates that are at least 'kUpdateStabilityParam' (@== 4k@) deep,
-  and then construct
-
-  > State
-  > { nextProtocolVersion    = cpuProtocolVersion
-  > , nextProtocolParameters = cpuProtocolParameters
-  > }
-
-  (with 'State' from "Cardano.Chain.Update.Validation.Interface.ProtocolVersionBump")
-  where 'cpuProtocolVersion'/'cpuProtocolParameters' are the version and
-  parameters from the update. This then ends up in the following callstack
-
-  > applyChainTick
-  > |
-  > \-- epochTransition
-  >     |
-  >     \-- registerEpoch
-  >         |
-  >         \-- tryBumpVersion
-
-  Now, if this is changing the major version of the protocol, then this actually
-  indicates the transition to Shelley, and the Byron 'applyChainTick' won't
-  actually happen. Instead, in 'singleEraTransition' we will report the
-  'EpochNo' of the transition as soon as it's @2k@ (not @4k@!) deep: in other
-  words, as soon as it is stable; at this point, the HFC will do the rest.
-
-  A slightly subtle point is that the Byron ledger does not record any
-  information about /past/ updates to the protocol parameters, and so if we
-  /were/ to ask the Byron ledger /after/ the update when the transition is
-  going to take place (did take place), it will say 'Nothing': transition not
-  yet known. In practice this won't matter, as it will have been translated to
-  a Shelley ledger at that point.
--------------------------------------------------------------------------------}
-
-byronTransition :: PartialLedgerConfig ByronBlock
-                -> Word16   -- ^ Shelley major protocol version
-                -> LedgerState ByronBlock
-                -> Maybe EpochNo
-byronTransition ByronPartialLedgerConfig{..} shelleyMajorVersion state =
-      takeAny
-    . mapMaybe isTransitionToShelley
-    . Byron.Inspect.protocolUpdates byronLedgerConfig
-    $ state
-  where
-    ByronTransitionInfo transitionInfo = byronLedgerTransition state
-
-    genesis = byronLedgerConfig
-    k       = CC.Genesis.gdK $ CC.Genesis.configGenesisData genesis
-
-    isTransitionToShelley :: Byron.Inspect.ProtocolUpdate -> Maybe EpochNo
-    isTransitionToShelley update = do
-        guard $ CC.Update.pvMajor version == shelleyMajorVersion
-        case Byron.Inspect.protocolUpdateState update of
-          Byron.Inspect.UpdateCandidate _becameCandidateSlotNo adoptedIn -> do
-            becameCandidateBlockNo <- Map.lookup version transitionInfo
-            guard $ isReallyStable becameCandidateBlockNo
-            return adoptedIn
-          Byron.Inspect.UpdateStableCandidate adoptedIn ->
-            -- If the Byron ledger thinks it's stable, it's _definitely_ stable
-            return adoptedIn
-          _otherwise ->
-            -- The proposal isn't yet a candidate, never mind a stable one
-            mzero
-      where
-        version :: CC.Update.ProtocolVersion
-        version = Byron.Inspect.protocolUpdateVersion update
-
-    -- Normally, stability in the ledger is defined in terms of slots, not
-    -- blocks. Byron considers the proposal to be stable after the slot is more
-    -- than @2k@ old. That is not wrong: after @2k@, the block indeed is stable.
-    --
-    -- Unfortunately, this means that the /conclusion about stability itself/
-    -- is /not/ stable: if we were to switch to a denser fork, we might change
-    -- our mind (on the sparse chain we thought the block was already stable,
-    -- but on the dense chain we conclude it is it not yet stable).
-    --
-    -- It is unclear at the moment if this presents a problem; the HFC assumes
-    -- monotonicity of timing info, in the sense that that any slot/time
-    -- conversions are either unknown or else not subject to rollback.
-    -- The problem sketched above might mean that we can go from "conversion
-    -- known" to "conversion unknown", but then when we go back again to
-    -- "conversion known", we /are/ guaranteed that we'd get the same answer.
-    --
-    -- Rather than trying to analyse this subtle problem, we instead base
-    -- stability on block numbers; after the block is `k` deep, we know for sure
-    -- that it is stable, and moreover, no matter which chain we switch to, that
-    -- will remain to be the case.
-    --
-    -- The Byron 'UpdateState' records the 'SlotNo' of the block in which the
-    -- proposal became a candidate (i.e., when the last required endorsement
-    -- came in). That doesn't tell us very much, we need to know the block
-    -- number; that's precisely what the 'ByronTransition' part of the Byron
-    -- state tells us.
-    isReallyStable :: BlockNo -> Bool
-    isReallyStable (BlockNo bno) = distance >= CC.unBlockCount k
-      where
-        distance :: Word64
-        distance = case byronLedgerTipBlockNo state of
-                     Origin                  -> bno + 1
-                     NotOrigin (BlockNo tip) -> tip - bno
-
-    -- We only expect a single proposal that updates to Shelley, but in case
-    -- there are multiple, any one will do
-    takeAny :: [a] -> Maybe a
-    takeAny = listToMaybe
-
-{-------------------------------------------------------------------------------
-  SingleEraBlock Byron
--------------------------------------------------------------------------------}
-
-instance SingleEraBlock ByronBlock where
-  singleEraTransition pcfg _eraParams _eraStart ledgerState =
-      case byronTriggerHardFork pcfg of
-        TriggerHardForkNotDuringThisExecution        -> Nothing
-        TriggerHardForkAtEpoch   epoch               -> Just epoch
-        TriggerHardForkAtVersion shelleyMajorVersion ->
-            byronTransition
-              pcfg
-              shelleyMajorVersion
-              ledgerState
-
-  singleEraInfo _ = SingleEraInfo {
-      singleEraName = "Byron"
-    }
-
-instance PBftCrypto bc => HasPartialConsensusConfig (PBft bc)
-  -- Use defaults
-
--- | When Byron is part of the hard-fork combinator, we use the partial ledger
--- config. Standalone Byron uses the regular ledger config. This means that
--- the partial ledger config is the perfect place to store the trigger
--- condition for the hard fork to Shelley, as we don't have to modify the
--- ledger config for standalone Byron.
-data ByronPartialLedgerConfig = ByronPartialLedgerConfig {
-      byronLedgerConfig    :: !(LedgerConfig ByronBlock)
-    , byronTriggerHardFork :: !TriggerHardFork
-    }
-  deriving (Generic, NoThunks)
-
-instance HasPartialLedgerConfig ByronBlock where
-
-  type PartialLedgerConfig ByronBlock = ByronPartialLedgerConfig
-
-  completeLedgerConfig _ _ = byronLedgerConfig
 
 {-------------------------------------------------------------------------------
   CanHardFork
@@ -474,8 +305,8 @@ translateChainDepStateByronToShelley TPraosConfig { tpraosParams } pbftState =
       SL.ChainDepState
         { SL.csProtocol = SL.PrtclState Map.empty nonce nonce
         , SL.csTickn    = SL.TicknState {
-              ticknStateEpochNonce    = nonce
-            , ticknStatePrevHashNonce = SL.NeutralNonce
+              SL.ticknStateEpochNonce    = nonce
+            , SL.ticknStatePrevHashNonce = SL.NeutralNonce
             }
           -- Overridden before used
         , SL.csLabNonce = SL.NeutralNonce
